Rename lib btc_wallet to bitcoin_wallet
