+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 
 members = [
-    "btc_htlc",
+    "bitcoin_htlc",
 	"jsonrpc",
 	"trading_service",
 	"exchange_service",
