+++ bitcoin_htlc/src/bitcoin_htlc.rs
