+++ bitcoin_htlc/src/lib.rs
@@ -4,7 +4,7 @@ extern crate bitcoin;
 extern crate common_types;
 extern crate hex;
 
-pub mod btc_htlc;
+pub mod bitcoin_htlc;
 
-pub use btc_htlc::*;
+pub use bitcoin_htlc::*;
 pub use common_types::secret;
