+++ trading_service/Cargo.toml
@@ -22,7 +22,7 @@ serde_derive = "1.0.51"
 serde_json = "1.0.17"
 bitcoin_rpc = { path = "../bitcoin_rpc" }
 bitcoin = { git = "https://github.com/tenx-tech/rust-bitcoin.git", branch = "bitcoincore-regtest" }
-bitcoin_htlc = { path = "../btc_htlc" }
+bitcoin_htlc = { path = "../bitcoin_htlc" }
 common_types = { path = "../common_types" }
 hex = "0.3.2"
 web3 = "0.3.0"
